Special handling for empty strings
The current implementation of the redisgraph-bulk-loader does
not handle empty strings as it doesn't differentiate between
an empty string and a property value that does not exist when
reading the csvfile.

The simplest solution here seemed to be setting property values
that should be `= ""` to `<<EMPTY>>` and having the bulk loader
make sure that it writes the a binary encoded empty string when
creating a new graph. All non existent property values will be treated
as NULL.